requirements.txt
flask
torch==2.4.0
torchvision==0.19.0
opencv-python
pillow==10.4.0
